+++ src/CAN/CanInterface.h
@@ -81,7 +81,7 @@ namespace CanInterface
 	GCodeResult SetRemotePressureAdvance(const CanDriversData<float>& data, const StringRef& reply) noexcept;
 	GCodeResult SetRemoteDriverStepsPerMmAndMicrostepping(const CanDriversData<StepsPerUnitAndMicrostepping>& data, const StringRef& reply) noexcept;
 #if SUPPORT_PHASE_STEPPING
-	GCodeResult SetRemoteStepMode(const CanDriversData<uint8_t>& data, const StringRef& reply) noexcept;
+	GCodeResult SetRemoteStepMode(const CanDriversData<uint16_t>& data, const StringRef& reply) noexcept;
 	GCodeResult SetRemotePhaseStepKv(const CanDriversData<float>& data, const StringRef& reply) noexcept;
 	GCodeResult SetRemotePhaseStepKa(const CanDriversData<float>& data, const StringRef& reply) noexcept;
 #endif
